correct the path to a specific view

+++ todolist_app/views.py
@@ -1,31 +1,41 @@
 from django.shortcuts import render, redirect
-from .models import TaskList
-from .forms import TaskForm
+from django.http import HttpResponse
+from todolist_app.models import TaskList
+from todolist_app.forms import TaskForm
 from django.contrib import messages
 from django.core.paginator import Paginator
+from django.contrib.auth.decorators import login_required
 
-# Create your views here.
 
+@login_required
 def todolist(request):
-    if request.method == 'POST':
+    if request.method == "POST":
         form = TaskForm(request.POST or None)
         if form.is_valid():
-            form.save()
-        messages.success(request, 'Task Added successfully!')
-        return redirect('todolist') 
-    else: ## Get request
-        all_tasks = TaskList.objects.all()
-        paginator = Paginator(all_tasks,3)
+            instance = form.save(commit=False)
+            instance.manage = request.user
+            instance.save()
+        messages.success(request,("New Task Added!"))
+        return redirect('todolist')
+    else: 
+        all_tasks = TaskList.objects.filter(manage=request.user)
+        paginator = Paginator(all_tasks, 10)
         page = request.GET.get('pg')
         all_tasks = paginator.get_page(page)
-        
-        return render(request, 'todolist.html',{'all_tasks':all_tasks})
 
-def delete_task(reques, task_id):
-    task = TaskList.objects.get(pk=task_id)  
-    task.delete()
+        return render(request, 'todolist.html', {'all_tasks': all_tasks})
+
+@login_required
+def delete_task(request, task_id):
+    task = TaskList.objects.get(pk=task_id)
+    if task.manage == request.user:
+        task.delete()
+    else:
+       messages.error(request,("Access Restricted, You Are Not Allowed."))
+
     return redirect('todolist')
 
+@login_required
 def edit_task(request, task_id):
     if request.method == "POST":
         task = TaskList.objects.get(pk=task_id)
@@ -38,7 +48,8 @@ def edit_task(request, task_id):
     else: 
         task_obj = TaskList.objects.get(pk=task_id)
         return render(request, 'edit.html', {'task_obj': task_obj})
-    
+
+@login_required
 def complete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
     if task.manage == request.user:
@@ -49,6 +60,7 @@ def complete_task(request, task_id):
 
     return redirect('todolist')
 
+@login_required
 def pending_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
     task.done = False
@@ -56,15 +68,20 @@ def pending_task(request, task_id):
 
     return redirect('todolist')
 
-  
-def contactus(request):
-    content = {
-        'todo_from' : 'Welcome to Contact Page'
-    }
-    return render(request, 'todolist.html',content)
+def index(request):
+    context = {
+        'index_text':"Welcome Index Page.",
+        }
+    return render(request, 'index.html', context)
 
-def aboutus(request):
-    content = {
-        'todo_from' : 'Welcome to About Page'
-    }
-    return render(request, 'todolist.html',content)
+def contact(request):
+    context = {
+        'contact_text':"Welcome Contact Page.",
+        }
+    return render(request, 'contact.html', context)
+
+def about(request):
+    context = {
+        'about_text':"Welcome About Page.",
+        }
+    return render(request, 'about.html', context)
